pyproject.toml
[tool.poetry]
name = "pyinverstadvisor"
version = "0.1"
description = "This system is an AI chatbot that analyzes a user's investment profile to recommend a personalized financial portfolio. Leveraging natural language processing, it is designed to allow anyone to easily ask investment-related questions and receive real-time financial information."
authors = [
    { name = "wonhyeok kim", email = "novatecoder@gmail.com" }
]
license = "Apache License 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-adk = "^1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"